REFACTOR: 소셜과 일반 인증용 패키지 구분
 - 일반 인증은 auth/local 하위, 소셜 인증은 auth/social 하위로 구분한다
 - 공통으로 쓸 클래스는 auth 하위에 남긴다
