+++ src/main/kotlin/com/fesi/flowit/auth/local/web/AuthController.kt
@@ -1,11 +1,11 @@
-package com.fesi.flowit.auth.web
+package com.fesi.flowit.auth.local.web
 
-import com.fesi.flowit.auth.service.AuthService
-import com.fesi.flowit.auth.service.dto.SignInDto
-import com.fesi.flowit.auth.web.request.RegenerateRequest
-import com.fesi.flowit.auth.web.request.SignInRequest
-import com.fesi.flowit.auth.web.response.RegenerateResponse
-import com.fesi.flowit.auth.web.response.SignInResponse
+import com.fesi.flowit.auth.local.service.AuthService
+import com.fesi.flowit.auth.local.service.dto.SignInDto
+import com.fesi.flowit.auth.local.web.request.RegenerateRequest
+import com.fesi.flowit.auth.local.web.request.SignInRequest
+import com.fesi.flowit.auth.local.web.response.RegenerateResponse
+import com.fesi.flowit.auth.local.web.response.SignInResponse
 import com.fesi.flowit.common.logging.loggerFor
 import com.fesi.flowit.common.response.ApiResponse
 import com.fesi.flowit.common.response.ApiResult
@@ -14,7 +14,6 @@ import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RestController
-import org.swyp.weddy.domain.auth.web.AuthApiSpec
 
 private val log = loggerFor<AuthController>()
